+++ hudi-client/hudi-spark-client/src/test/java/org/apache/hudi/table/TestHoodieMergeOnReadTable.java
@@ -26,7 +26,6 @@ import org.apache.hudi.common.model.FileSlice;
 import org.apache.hudi.common.model.HoodieBaseFile;
 import org.apache.hudi.common.model.HoodieCommitMetadata;
 import org.apache.hudi.common.model.HoodieFileFormat;
-import org.apache.hudi.common.model.HoodieFileGroup;
 import org.apache.hudi.common.model.HoodieRecord;
 import org.apache.hudi.common.model.HoodieTableType;
 import org.apache.hudi.common.model.HoodieWriteStat;
@@ -37,7 +36,6 @@ import org.apache.hudi.common.table.timeline.HoodieInstant;
 import org.apache.hudi.common.table.timeline.HoodieInstant.State;
 import org.apache.hudi.common.table.timeline.HoodieTimeline;
 import org.apache.hudi.common.table.view.FileSystemViewStorageConfig;
-import org.apache.hudi.common.table.view.HoodieTableFileSystemView;
 import org.apache.hudi.common.table.view.SyncableFileSystemView;
 import org.apache.hudi.common.table.view.TableFileSystemView.BaseFileOnlyView;
 import org.apache.hudi.common.table.view.TableFileSystemView.SliceView;
@@ -77,7 +75,6 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -149,409 +146,6 @@ public class TestHoodieMergeOnReadTable extends HoodieClientTestHarness {
       assertEquals(String.valueOf(2 * records.size()),
           allWriteStatusMergedMetadataMap.get("InputRecordCount_1506582000"));
     }
-  }
-
-  private void testCOWToMORConvertedTableRollback(Boolean rollbackUsingMarkers) throws Exception {
-    // Set TableType to COW
-    HoodieTestUtils.init(hadoopConf, basePath, HoodieTableType.COPY_ON_WRITE);
-
-    HoodieWriteConfig cfg = getConfig(false, rollbackUsingMarkers);
-    try (SparkRDDWriteClient client = getHoodieWriteClient(cfg);) {
-
-      /**
-       * Write 1 (only inserts)
-       */
-      String newCommitTime = "001";
-      client.startCommitWithTime(newCommitTime);
-
-      List<HoodieRecord> records = dataGen.generateInserts(newCommitTime, 200);
-      JavaRDD<HoodieRecord> writeRecords = jsc.parallelize(records, 1);
-
-      List<WriteStatus> statuses = client.upsert(writeRecords, newCommitTime).collect();
-      // verify there are no errors
-      assertNoWriteErrors(statuses);
-      client.commit(newCommitTime, jsc.parallelize(statuses));
-
-      metaClient = getHoodieMetaClient(hadoopConf, cfg.getBasePath());
-      Option<HoodieInstant> commit = metaClient.getActiveTimeline().getCommitTimeline().firstInstant();
-      assertTrue(commit.isPresent());
-      assertEquals("001", commit.get().getTimestamp(), "commit should be 001");
-
-      /**
-       * Write 2 (updates)
-       */
-      newCommitTime = "002";
-      client.startCommitWithTime(newCommitTime);
-
-      records = dataGen.generateUpdates(newCommitTime, records);
-
-      statuses = client.upsert(jsc.parallelize(records, 1), newCommitTime).collect();
-      // Verify there are no errors
-      assertNoWriteErrors(statuses);
-
-      // Set TableType to MOR
-      HoodieTestUtils.init(hadoopConf, basePath, HoodieTableType.MERGE_ON_READ);
-
-      // rollback a COW commit when TableType is MOR
-      client.rollback(newCommitTime);
-
-      metaClient = HoodieTableMetaClient.reload(metaClient);
-      HoodieTable hoodieTable = HoodieSparkTable.create(cfg, context, metaClient);
-      FileStatus[] allFiles = listAllBaseFilesInPath(hoodieTable);
-      tableView = getHoodieTableFileSystemView(metaClient, hoodieTable.getCompletedCommitsTimeline(), allFiles);
-
-      final String absentCommit = newCommitTime;
-      assertFalse(tableView.getLatestBaseFiles().anyMatch(file -> absentCommit.equals(file.getCommitTime())));
-    }
-  }
-
-  @Test
-  public void testCOWToMORConvertedTableRollbackUsingFileList() throws Exception {
-    testCOWToMORConvertedTableRollback(false);
-  }
-
-  @Test
-  public void testCOWToMORConvertedTableRollbackUsingMarkers() throws Exception {
-    testCOWToMORConvertedTableRollback(true);
-  }
-
-  private void testRollbackWithDeltaAndCompactionCommit(Boolean rollbackUsingMarkers, boolean populateMetaFields) throws Exception {
-    HoodieWriteConfig.Builder cfgBuilder = getConfigBuilder(false, rollbackUsingMarkers, IndexType.SIMPLE);
-    addConfigsForPopulateMetaFields(cfgBuilder, populateMetaFields);
-    HoodieWriteConfig cfg = cfgBuilder.build();
-
-    try (SparkRDDWriteClient client = getHoodieWriteClient(cfg);) {
-
-      // Test delta commit rollback
-      /**
-       * Write 1 (only inserts)
-       */
-      String newCommitTime = "001";
-      client.startCommitWithTime(newCommitTime);
-
-      List<HoodieRecord> records = dataGen.generateInserts(newCommitTime, 200);
-      JavaRDD<HoodieRecord> writeRecords = jsc.parallelize(records, 1);
-
-      JavaRDD<WriteStatus> writeStatusJavaRDD = client.upsert(writeRecords, newCommitTime);
-      client.commit(newCommitTime, writeStatusJavaRDD);
-      List<WriteStatus> statuses = writeStatusJavaRDD.collect();
-      assertNoWriteErrors(statuses);
-
-      metaClient = getHoodieMetaClient(hadoopConf, cfg.getBasePath());
-      HoodieTable hoodieTable = HoodieSparkTable.create(cfg, context, metaClient);
-
-      Option<HoodieInstant> deltaCommit = metaClient.getActiveTimeline().getDeltaCommitTimeline().firstInstant();
-      assertTrue(deltaCommit.isPresent());
-      assertEquals("001", deltaCommit.get().getTimestamp(), "Delta commit should be 001");
-
-      Option<HoodieInstant> commit = metaClient.getActiveTimeline().getCommitTimeline().firstInstant();
-      assertFalse(commit.isPresent());
-
-      FileStatus[] allFiles = listAllBaseFilesInPath(hoodieTable);
-      tableView =
-          getHoodieTableFileSystemView(metaClient, metaClient.getCommitTimeline().filterCompletedInstants(), allFiles);
-      Stream<HoodieBaseFile> dataFilesToRead = tableView.getLatestBaseFiles();
-      assertTrue(!dataFilesToRead.findAny().isPresent());
-
-      tableView = getHoodieTableFileSystemView(metaClient, hoodieTable.getCompletedCommitsTimeline(), allFiles);
-      dataFilesToRead = tableView.getLatestBaseFiles();
-      assertTrue(dataFilesToRead.findAny().isPresent(),
-          "should list the base files we wrote in the delta commit");
-
-      /**
-       * Write 2 (inserts + updates - testing failed delta commit)
-       */
-      final String commitTime1 = "002";
-      // WriteClient with custom config (disable small file handling)
-      try (SparkRDDWriteClient secondClient = getHoodieWriteClient(getHoodieWriteConfigWithSmallFileHandlingOff(false));) {
-        secondClient.startCommitWithTime(commitTime1);
-
-        List<HoodieRecord> copyOfRecords = new ArrayList<>(records);
-        copyOfRecords = dataGen.generateUpdates(commitTime1, copyOfRecords);
-        copyOfRecords.addAll(dataGen.generateInserts(commitTime1, 200));
-
-        List<String> dataFiles = tableView.getLatestBaseFiles().map(HoodieBaseFile::getPath).collect(Collectors.toList());
-        List<GenericRecord> recordsRead = HoodieMergeOnReadTestUtils.getRecordsUsingInputFormat(hadoopConf, dataFiles,
-            basePath);
-        assertEquals(200, recordsRead.size());
-
-        statuses = secondClient.upsert(jsc.parallelize(copyOfRecords, 1), commitTime1).collect();
-        // Verify there are no errors
-        assertNoWriteErrors(statuses);
-
-        // Test failed delta commit rollback
-        secondClient.rollback(commitTime1);
-        allFiles = listAllBaseFilesInPath(hoodieTable);
-        // After rollback, there should be no base file with the failed commit time
-        List<String> remainingFiles = Arrays.stream(allFiles).filter(file -> file.getPath().getName()
-            .contains(commitTime1)).map(fileStatus -> fileStatus.getPath().toString()).collect(Collectors.toList());
-        assertEquals(0, remainingFiles.size(), "There files should have been rolled-back "
-            + "when rolling back commit " + commitTime1 + " but are still remaining. Files: " + remainingFiles);
-        dataFiles = tableView.getLatestBaseFiles().map(HoodieBaseFile::getPath).collect(Collectors.toList());
-        recordsRead = HoodieMergeOnReadTestUtils.getRecordsUsingInputFormat(hadoopConf, dataFiles, basePath);
-        assertEquals(200, recordsRead.size());
-      }
-
-      /**
-       * Write 3 (inserts + updates - testing successful delta commit)
-       */
-      final String commitTime2 = "002";
-      try (SparkRDDWriteClient thirdClient = getHoodieWriteClient(cfg);) {
-        thirdClient.startCommitWithTime(commitTime2);
-
-        List<HoodieRecord> copyOfRecords = new ArrayList<>(records);
-        copyOfRecords = dataGen.generateUpdates(commitTime2, copyOfRecords);
-        copyOfRecords.addAll(dataGen.generateInserts(commitTime2, 200));
-
-        List<String> dataFiles = tableView.getLatestBaseFiles().map(HoodieBaseFile::getPath).collect(Collectors.toList());
-        List<GenericRecord> recordsRead = HoodieMergeOnReadTestUtils.getRecordsUsingInputFormat(hadoopConf, dataFiles,
-            basePath);
-        assertEquals(200, recordsRead.size());
-
-        writeRecords = jsc.parallelize(copyOfRecords, 1);
-        writeStatusJavaRDD = thirdClient.upsert(writeRecords, commitTime2);
-        statuses = writeStatusJavaRDD.collect();
-        // Verify there are no errors
-        assertNoWriteErrors(statuses);
-
-        // Test successful delta commit rollback
-        thirdClient.rollback(commitTime2);
-        allFiles = listAllBaseFilesInPath(hoodieTable);
-        // After rollback, there should be no base file with the failed commit time
-        assertEquals(0, Arrays.stream(allFiles)
-            .filter(file -> file.getPath().getName().contains(commitTime2)).count());
-
-        metaClient = HoodieTableMetaClient.reload(metaClient);
-        hoodieTable = HoodieSparkTable.create(cfg, context, metaClient);
-        tableView = getHoodieTableFileSystemView(metaClient, hoodieTable.getCompletedCommitsTimeline(), allFiles);
-        dataFiles = tableView.getLatestBaseFiles().map(HoodieBaseFile::getPath).collect(Collectors.toList());
-        recordsRead = HoodieMergeOnReadTestUtils.getRecordsUsingInputFormat(hadoopConf, dataFiles, basePath);
-        // check that the number of records read is still correct after rollback operation
-        assertEquals(200, recordsRead.size());
-
-        // Test compaction commit rollback
-        /**
-         * Write 4 (updates)
-         */
-        newCommitTime = "003";
-        thirdClient.startCommitWithTime(newCommitTime);
-
-        writeStatusJavaRDD = thirdClient.upsert(writeRecords, newCommitTime);
-        statuses = writeStatusJavaRDD.collect();
-        thirdClient.commit(newCommitTime, writeStatusJavaRDD);
-        // Verify there are no errors
-        assertNoWriteErrors(statuses);
-
-        metaClient = HoodieTableMetaClient.reload(metaClient);
-
-        String compactionInstantTime = thirdClient.scheduleCompaction(Option.empty()).get().toString();
-        thirdClient.compact(compactionInstantTime);
-
-        allFiles = listAllBaseFilesInPath(hoodieTable);
-        metaClient = HoodieTableMetaClient.reload(metaClient);
-        tableView = getHoodieTableFileSystemView(metaClient, metaClient.getCommitsTimeline(), allFiles);
-
-        final String compactedCommitTime = metaClient.getActiveTimeline().reload().lastInstant().get().getTimestamp();
-        assertTrue(Arrays.stream(listAllBaseFilesInPath(hoodieTable))
-                .anyMatch(file -> compactedCommitTime.equals(new HoodieBaseFile(file).getCommitTime())));
-        thirdClient.rollbackInflightCompaction(new HoodieInstant(State.INFLIGHT, HoodieTimeline.COMPACTION_ACTION, compactedCommitTime),
-                hoodieTable);
-        allFiles = listAllBaseFilesInPath(hoodieTable);
-        metaClient = HoodieTableMetaClient.reload(metaClient);
-        tableView = getHoodieTableFileSystemView(metaClient, metaClient.getCommitsTimeline(), allFiles);
-
-        assertFalse(tableView.getLatestBaseFiles().anyMatch(file -> compactedCommitTime.equals(file.getCommitTime())));
-      }
-    }
-  }
-
-  @ParameterizedTest
-  @MethodSource("populateMetaFieldsParams")
-  public void testRollbackWithDeltaAndCompactionCommitUsingFileList(boolean populateMetaFields) throws Exception {
-    clean();
-    init(HoodieTableConfig.BASE_FILE_FORMAT.defaultValue(), populateMetaFields);
-    testRollbackWithDeltaAndCompactionCommit(false, populateMetaFields);
-  }
-
-  @ParameterizedTest
-  @MethodSource("populateMetaFieldsParams")
-  public void testRollbackWithDeltaAndCompactionCommitUsingMarkers(boolean populateMetaFields) throws Exception {
-    clean();
-    init(HoodieTableConfig.BASE_FILE_FORMAT.defaultValue(), populateMetaFields);
-    testRollbackWithDeltaAndCompactionCommit(true, populateMetaFields);
-  }
-
-  @ParameterizedTest
-  @MethodSource("populateMetaFieldsParams")
-  public void testMultiRollbackWithDeltaAndCompactionCommit(boolean populateMetaFields) throws Exception {
-    clean();
-    init(HoodieTableConfig.BASE_FILE_FORMAT.defaultValue(), populateMetaFields);
-    HoodieWriteConfig.Builder cfgBuilder = getConfigBuilder(false);
-    addConfigsForPopulateMetaFields(cfgBuilder, populateMetaFields);
-    HoodieWriteConfig cfg = cfgBuilder.build();
-
-    try (final SparkRDDWriteClient client = getHoodieWriteClient(cfg);) {
-      /**
-       * Write 1 (only inserts)
-       */
-      String newCommitTime = "001";
-      client.startCommitWithTime(newCommitTime);
-
-      List<HoodieRecord> records = dataGen.generateInserts(newCommitTime, 200);
-      JavaRDD<HoodieRecord> writeRecords = jsc.parallelize(records, 1);
-
-      JavaRDD<WriteStatus> writeStatusJavaRDD = client.upsert(writeRecords, newCommitTime);
-      client.commit(newCommitTime, writeStatusJavaRDD);
-      List<WriteStatus> statuses = writeStatusJavaRDD.collect();
-      assertNoWriteErrors(statuses);
-
-      metaClient = getHoodieMetaClient(hadoopConf, cfg.getBasePath());
-      HoodieTable hoodieTable = HoodieSparkTable.create(cfg, context, metaClient);
-
-      Option<HoodieInstant> deltaCommit = metaClient.getActiveTimeline().getDeltaCommitTimeline().firstInstant();
-      assertTrue(deltaCommit.isPresent());
-      assertEquals("001", deltaCommit.get().getTimestamp(), "Delta commit should be 001");
-
-      Option<HoodieInstant> commit = metaClient.getActiveTimeline().getCommitTimeline().firstInstant();
-      assertFalse(commit.isPresent());
-
-      FileStatus[] allFiles = listAllBaseFilesInPath(hoodieTable);
-      tableView = getHoodieTableFileSystemView(metaClient, metaClient.getCommitTimeline().filterCompletedInstants(), allFiles);
-      Stream<HoodieBaseFile> dataFilesToRead = tableView.getLatestBaseFiles();
-      assertFalse(dataFilesToRead.findAny().isPresent());
-
-      tableView = getHoodieTableFileSystemView(metaClient, hoodieTable.getCompletedCommitsTimeline(), allFiles);
-      dataFilesToRead = tableView.getLatestBaseFiles();
-      assertTrue(dataFilesToRead.findAny().isPresent(),
-          "Should list the base files we wrote in the delta commit");
-
-      /**
-       * Write 2 (inserts + updates)
-       */
-      newCommitTime = "002";
-      // WriteClient with custom config (disable small file handling)
-      SparkRDDWriteClient nClient = getHoodieWriteClient(getHoodieWriteConfigWithSmallFileHandlingOff(populateMetaFields));
-      nClient.startCommitWithTime(newCommitTime);
-
-      List<HoodieRecord> copyOfRecords = new ArrayList<>(records);
-      copyOfRecords = dataGen.generateUpdates(newCommitTime, copyOfRecords);
-      copyOfRecords.addAll(dataGen.generateInserts(newCommitTime, 200));
-
-      List<String> dataFiles = tableView.getLatestBaseFiles().map(hf -> hf.getPath()).collect(Collectors.toList());
-      List<GenericRecord> recordsRead = HoodieMergeOnReadTestUtils.getRecordsUsingInputFormat(hadoopConf, dataFiles,
-          basePath);
-      assertEquals(200, recordsRead.size());
-
-      statuses = nClient.upsert(jsc.parallelize(copyOfRecords, 1), newCommitTime).collect();
-      // Verify there are no errors
-      assertNoWriteErrors(statuses);
-      nClient.commit(newCommitTime, writeStatusJavaRDD);
-      copyOfRecords.clear();
-
-      // Schedule a compaction
-      /**
-       * Write 3 (inserts + updates)
-       */
-      newCommitTime = "003";
-      client.startCommitWithTime(newCommitTime);
-
-      List<HoodieRecord> newInserts = dataGen.generateInserts(newCommitTime, 100);
-      records = dataGen.generateUpdates(newCommitTime, records);
-      records.addAll(newInserts);
-      writeRecords = jsc.parallelize(records, 1);
-
-      writeStatusJavaRDD = client.upsert(writeRecords, newCommitTime);
-      client.commit(newCommitTime, writeStatusJavaRDD);
-      statuses = writeStatusJavaRDD.collect();
-      // Verify there are no errors
-      assertNoWriteErrors(statuses);
-
-      metaClient = HoodieTableMetaClient.reload(metaClient);
-
-      String compactionInstantTime = "004";
-      client.scheduleCompactionAtInstant(compactionInstantTime, Option.empty());
-
-      // Compaction commit
-      /**
-       * Write 4 (updates)
-       */
-      newCommitTime = "005";
-      client.startCommitWithTime(newCommitTime);
-
-      records = dataGen.generateUpdates(newCommitTime, records);
-      writeRecords = jsc.parallelize(records, 1);
-
-      writeStatusJavaRDD = client.upsert(writeRecords, newCommitTime);
-      client.commit(newCommitTime, writeStatusJavaRDD);
-      statuses = writeStatusJavaRDD.collect();
-      // Verify there are no errors
-      assertNoWriteErrors(statuses);
-
-      metaClient = HoodieTableMetaClient.reload(metaClient);
-
-      compactionInstantTime = "006";
-      client.scheduleCompactionAtInstant(compactionInstantTime, Option.empty());
-      JavaRDD<WriteStatus> ws = (JavaRDD<WriteStatus>) client.compact(compactionInstantTime);
-      client.commitCompaction(compactionInstantTime, ws, Option.empty());
-
-      allFiles = listAllBaseFilesInPath(hoodieTable);
-      metaClient = HoodieTableMetaClient.reload(metaClient);
-      tableView = getHoodieTableFileSystemView(metaClient, metaClient.getCommitsTimeline(), allFiles);
-
-      final String compactedCommitTime =
-          metaClient.getActiveTimeline().reload().getCommitsTimeline().lastInstant().get().getTimestamp();
-
-      assertTrue(tableView.getLatestBaseFiles().anyMatch(file -> compactedCommitTime.equals(file.getCommitTime())));
-
-      /**
-       * Write 5 (updates)
-       */
-      newCommitTime = "007";
-      client.startCommitWithTime(newCommitTime);
-      copyOfRecords = new ArrayList<>(records);
-      copyOfRecords = dataGen.generateUpdates(newCommitTime, copyOfRecords);
-      copyOfRecords.addAll(dataGen.generateInserts(newCommitTime, 200));
-
-      statuses = client.upsert(jsc.parallelize(copyOfRecords, 1), newCommitTime).collect();
-      // Verify there are no errors
-      assertNoWriteErrors(statuses);
-      client.commit(newCommitTime, writeStatusJavaRDD);
-      copyOfRecords.clear();
-
-      // Rollback latest commit first
-      client.restoreToInstant("000");
-
-      metaClient = HoodieTableMetaClient.reload(metaClient);
-      allFiles = listAllBaseFilesInPath(hoodieTable);
-      tableView = getHoodieTableFileSystemView(metaClient, metaClient.getCommitTimeline().filterCompletedInstants(), allFiles);
-      dataFilesToRead = tableView.getLatestBaseFiles();
-      assertFalse(dataFilesToRead.findAny().isPresent());
-      SliceView rtView = getHoodieTableFileSystemView(metaClient, metaClient.getCommitTimeline().filterCompletedInstants(), allFiles);
-      List<HoodieFileGroup> fileGroups =
-          ((HoodieTableFileSystemView) rtView).getAllFileGroups().collect(Collectors.toList());
-      assertTrue(fileGroups.isEmpty());
-
-      // make sure there are no log files remaining
-      assertEquals(0L, ((HoodieTableFileSystemView) rtView).getAllFileGroups()
-          .filter(fileGroup -> fileGroup.getAllRawFileSlices().noneMatch(f -> f.getLogFiles().count() == 0))
-          .count());
-
-    }
-  }
-
-  protected HoodieWriteConfig getHoodieWriteConfigWithSmallFileHandlingOff(boolean populateMetaFields) {
-    HoodieWriteConfig.Builder cfgBuilder = HoodieWriteConfig.newBuilder().withPath(basePath).withSchema(TRIP_EXAMPLE_SCHEMA).withParallelism(2, 2)
-        .withDeleteParallelism(2)
-        .withAutoCommit(false)
-        .withCompactionConfig(HoodieCompactionConfig.newBuilder().compactionSmallFileSize(1024)
-            .withInlineCompaction(false).withMaxNumDeltaCommitsBeforeCompaction(1).build())
-        .withEmbeddedTimelineServerEnabled(true)
-        .withStorageConfig(HoodieStorageConfig.newBuilder().hfileMaxFileSize(1024).parquetMaxFileSize(1024).build()).forTable("test-trip-table");
-
-    if (!populateMetaFields) {
-      addConfigsForPopulateMetaFields(cfgBuilder, false);
-    }
-    return cfgBuilder.build();
   }
 
   @ParameterizedTest
